`Assert` changes: renamed to `MyAssert`, moved to new file, removed unused methods
